cleaning up cpp
